Name the header file "test.h", not "test.h++"
Aargh, this took forever to find!  It turns out that Chisel's
generates test harnesses that use "test.h", which was causing those
files to be built against the Chisel version of the header file (as
opposed to the LLVM version).  This was obviously bad, as the object
layout is completely different.

This should be fixed now...