Add a shallow_node class
Now we've got wide_nodes, which contain narrow_nodes, which contain
shallow_nodes.  The idea is that this deals with the three levels of
mapping: wide are uncanstrained, narrow fit in width, and shallow fit
in depth (and width).

This should let me transparently handle memories that are both too
wide and too deep to fit on a DREAMER tile.