Add a version of DES without bit shuffles
This test "works", which means it computes the same values as the C++
emulator does.  It doesn't actually do DES, but it's a somewhat
similar (though a bunch smaller) computation.

Hopefully this means that all I need to do is get the bit shuffles
working and then DES will pass.