Use "flo-llvm" (not "flo-llvm-c++")
I should have been using this wrapper all along!  This way I'm
somewhat isolated from the fact that there's a whole bunch of
compiler/linker magic going on inside the LLVM toolchain.

Note that this requires quite a few changes to the test harness: I can
no longer just symlink a bunch of stuff around but now instead
actually match the Chisel generated names explicitly.  This is
probably for the best (and should be ported back to flo-llvm) -- the
other way is super fragile!